requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
anthropic>=0.40.0
